backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.42.0
aiosqlite==0.22.1
pydantic==2.12.5
pydantic-settings==2.13.1
openai==2.29.0
chromadb==1.5.5
pyyaml==6.0.3
cryptography==46.0.5
python-jose[cryptography]==3.5.0
python-multipart==0.0.22
python-dotenv==1.1.0
httpx==0.28.1
passlib[bcrypt]==1.8.0bcrypt==4.2.0
# Testes
pytest==8.3.5
pytest-asyncio==0.24.0
